pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "solus"
version = "0.5.0"
description = "Local-first AI workflow engine for chaining inputs, transforms, and local LLM steps."
readme = "README.md"
authors = [
  { name = "Steven Foerster", email = "sfoerster@gmail.com" },
]
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0,<3",
  "PyYAML>=6.0,<7",
  "defusedxml>=0.7,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9",
  "pytest-cov>=5.0,<7",
  "ruff>=0.4,<1",
  "mypy>=1.10,<2",
  "types-requests>=2.31,<3",
]
pdf = [
  "pypdf>=4.0,<6",
]
ocr = [
  "pytesseract>=0.3.10,<1",
  "Pillow>=10.0,<12",
]
s3 = [
  "boto3>=1.34,<2",
]
vector = [
  "chromadb>=0.4,<1",
]
oidc = [
  "PyJWT>=2.8,<3",
  "cryptography>=41.0,<46",
]
encrypt = [
  "pysqlcipher3>=1.2,<2",
]

[project.scripts]
solus = "solus.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=solus --cov-report=term-missing --cov-report=html"

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
  "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["F"]

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
  "yaml",
  "defusedxml",
  "defusedxml.*",
  "jwt",
  "pytesseract",
  "PIL",
  "PIL.*",
  "pypdf",
  "chromadb",
  "boto3",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "solus.triggers.loader",
  "solus.modules.transform.ocr",
  "solus.modules.transform.metadata_extract",
  "solus.modules.output.vector_store",
  "solus.modules.input.parse_pdf",
  "solus.modules.input.s3_watcher",
]
disable_error_code = ["import-untyped", "import-not-found"]
